Index zendesk content
