+++ requirements.txt
@@ -3,6 +3,8 @@ atlassian_python_api==3.32.1
 beautifulsoup4==4.11.1
 html2text==2020.1.16
 nltk==3.8
+numpy==1.24.1
 openai==0.25.0
 pandas==1.5.2
+requests==2.28.1
 transformers==4.25.1
